+++ TSS.Java/src/tss/tpm/TPMS_RSA_PARMS.java
@@ -72,7 +72,7 @@ public class TPMS_RSA_PARMS extends TpmStructure implements TPMU_PUBLIC_PARMS
     {
         symmetric = _symmetric;
         scheme = _scheme;
-        keyBits = (short)_keyBits;
+        keyBits = _keyBits;
         exponent = _exponent;
     }
